+++ index.es6.js
@@ -25,6 +25,8 @@ from './source/phone'
 export
 {
 	default as country,
+	country_codes,
+	country_codes as countryCodes,
 	country_from_locale,
 	country_from_locale as countryFromLocale
 }